# src/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This api call data from youtube anf table",
        "title": "Dataset Collection API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "10.201.252.7:30014",
    "paths": {
...
        "/dataset/subtitle/{subtitleTagId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dataset subtitle"
                ],
                "summary": "get this subtitleTagId detail list",
                "operationId": "query-subtitle-by-subtitle-tagId-handler",
                "parameters": [
                    {
                        "type": "string",
                        "description": "subtitleTag id",
                        "name": "subtitleTagId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "We need subtitleTag ID!!",
                        "schema": {
                            "$ref": "#/definitions/string"
                        }
                    }
                }
            }
        },
...
        "/dataset/youtubeUrl/subtitle/{subtitleTagId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dataset subtitle"
                ],
                "summary": "download dataset by subtitleTagId",
                "operationId": "url-to-download-subtitle-tag",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Subtitle Tag ID",
                        "name": "subtitleTagId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "We need Subtitle Tag ID!!",
                        "schema": {
                            "$ref": "#/definitions/string"
                        }
                    }
                }
            }
        },
        "/dataset/youtubeUrl/subtitleById/{subtitleId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dataset subtitle"
                ],
                "summary": "download video by subtitleId",
                "operationId": "url-to-download-subtitle-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Subtitle ID",
                        "name": "subtitleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "We need Subtitle ID!!",
                        "schema": {
                            "$ref": "#/definitions/string"
                        }
                    }
                }
            }
        },
...
    }
}